+++ macrosynergy/visuals/table.py
@@ -4,7 +4,7 @@ import seaborn as sns
 import matplotlib.pyplot as plt
 
 from macrosynergy.signal.signal_return import SignalsReturns
-from macrosynergy.management.simulate_quantamental_data import make_qdf
+from macrosynergy.management.simulate import make_qdf
 
 
 def view_table(
